change type

+++ src/main/java/com/yeapkl/repository/PaymentDetailsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaymentDetailsRepository extends CrudRepository<PaymentDetails,Integer> {
+public interface PaymentDetailsRepository extends CrudRepository<PaymentDetails,Long> {
 }
